Separado o appbar e o Navegationbar
