+++ REPRESENTANTE/pacsoft_representante/lib/Components/Navegation.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+class Navigation extends StatefulWidget implements PreferredSizeWidget{
+  @override
+  State<StatefulWidget> createState() {
+    return baixo();
+  }
+
+  @override
+  // TODO: implement preferredSize
+  Size get preferredSize => throw UnimplementedError();
+ 
+  
+}
+
+class baixo extends State<Navigation>{
+  @override
+  Widget build(BuildContext context) {
+    return NavigationBar( //BARRA DE ESCOLHA
+        destinations: [
+        NavigationDestination(icon: Image.asset('assets/images/inicio.png',width: 40), label: 'Inicio'),
+        NavigationDestination(icon: Image.asset('assets/images/cliente.png',width: 40), label: 'Clientes'),
+        NavigationDestination(icon: Image.asset('assets/images/pedidos.png',width: 40), label: 'Pedidos'),
+        NavigationDestination(icon: Image.asset('assets/images/relatorio.png',width: 40), label: 'Relatórios'),
+      ]) ;
+  }
+  
+
+}
