+++ REPRESENTANTE/pacsoft_representante/lib/HOME/PG_INICIAL.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:pacsoft_representante/Components/Appbar.dart';
+import 'package:pacsoft_representante/Components/Navegation.dart';
 
 class PgInicial  extends StatefulWidget{
   @override
@@ -14,22 +16,7 @@ class pginicialstage extends State<PgInicial>{
   Widget build(BuildContext context) {
     // TODO: implement build
     return Scaffold(
-      appBar: AppBar( //BARRA SUPERIOR
-        titleSpacing: 30,
-        toolbarHeight: 100,
-        title: Image.asset('assets/images/logo.png', height: 60),
-        actions: <Widget>[
-          Padding(padding: EdgeInsets.only(right: 30),
-          child:  SizedBox(
-            child: IconButton(onPressed:(){
-
-            }, icon: Icon(Icons.account_circle_outlined,size: 40,)),
-          )
-        )
-        ],
-        backgroundColor: Colors.green[200],
-        elevation: 0,
-      ),
+      appBar: barra(height: 100),
 
       body: Container( //MENU DE ATALHO E PESQUISA
 
@@ -162,14 +149,8 @@ class pginicialstage extends State<PgInicial>{
           ],
       ),
       ),
-      bottomNavigationBar: NavigationBar( //BARRA DE ESCOLHA
-        destinations: [
-        NavigationDestination(icon: Image.asset('assets/images/inicio.png',width: 40), label: 'Inicio'),
-        NavigationDestination(icon: Image.asset('assets/images/cliente.png',width: 40), label: 'Clientes'),
-        NavigationDestination(icon: Image.asset('assets/images/pedidos.png',width: 40), label: 'Pedidos'),
-        NavigationDestination(icon: Image.asset('assets/images/relatorio.png',width: 40), label: 'Relatórios'),
-      ]) ,
-
+      
+      bottomNavigationBar: Navigation()
     );
   }
